Revert "fill holes of north and south pole" and related commits
Reverted the commits in the following order:

```
d4c4c91 - bug fix of dem tile (north & south pole)
8a00828 - bug fix of Globe.fileRayDistance: wrong position was retured
          when north or south pole was enabled
9e71ff0 - bug fix of picking globe
b573a02 - fill holes of north and south pole
```

The following files conflicted with revert of b573a02, so they were
resolved properly. The commits in parentheses are the commits that are
believed to have caused the conflict.

- mapray/src/FlakeMesh.js (9129132)
- mapray/src/Globe.ts (768b74b)
- mapray/src/Viewer.ts (805f837, 415001b, e6f0f9c)